Small fixes and threshold connection
Fixed gui argument being "GUI," now "--gui."
Moved all controller functionality to MainUI class from MainWindowUI.
Connected rank threshold texts to main body.
