+++ .json
@@ -1,0 +1,1 @@
+{"baseWord": ["r", "n", "b", "i", "m", "e", "o"], "foundWords": ["brine", "bromine"], "playerPoints": 19, "requiredLetter": "r", "maxPoints": 729}
