+++ BigPythonEnergy/MainCLI.py
@@ -147,7 +147,7 @@ def randomWord():
     startGame(newPuzzle)
 
 # Start script to clear the command line so the user just sees the instructions.
-if str(len(sys.argv) > 1 and sys.argv[1]) == "GUI":
+if str(len(sys.argv) > 1 and sys.argv[1]) == "--gui":
     exec(open("BigPythonEnergy/MainUI.py").read())
 print('--------------------------------')
 print('Welcome to Spelling Bee!')
