+++ love.json
@@ -1,0 +1,1 @@
+{"baseWord": ["b", "r", "o", "m", "n", "i", "e"], "foundWords": ["miner", "bromine"], "playerPoints": 19, "requiredLetter": "i", "maxPoints": 557}
